+++ test/Atc.Wpf.Components.Tests/GlobalUsings.cs
@@ -4,6 +4,7 @@ global using System.Windows.Controls;
 global using System.Windows.Data;
 
 global using Atc.Wpf.Components.Flyouts;
+global using Atc.Wpf.Components.Selectors;
 global using Atc.Wpf.Components.ValueConverters;
 global using Atc.Wpf.Components.Zoom;